refactor: nouveaux composants Simulateur
