+++ site/source/components/Simulateur/Boutons/BoutonDétail.tsx
