+++ site/source/components/Simulateur/Boutons/BoutonReset.tsx
